refact: ui package, and syncer interface
