+++ cmd/tt/ui/sidebar.go
@@ -1,0 +1,1 @@
+package ui
